fix: split session import functional test for backend-size limit
